Add the 'topic_is_web' column to the `topics` table
We must be able to choose which topics cannot participate in the formation of the catalog (sites, programs, etc.). Which can only take part in the formation of navigation through posts.

While the central catalog page is usually configured manually (config) to provide the best choice, internal themes are shown only due to their characteristics.