Adding a table `facets_matching`
The first step is to drop the `facet_related` column in the `facets` table. To connect trees among themselves, a more flexible system is to use a junction table than to write to a column separated by commas. We can use indexes and not break consistency in the database.

Upgrade to HLEB v1.5.79 (+ log rotation)